backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM (MariaDB)
sqlalchemy>=2.0.0
alembic>=1.13.0
mariadb>=1.1.0
pymysql>=1.1.0
cryptography>=42.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin compatible bcrypt version for passlib
python-multipart>=0.0.6

# Redis for caching
redis>=5.0.0
hiredis>=2.2.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Content processing and security
bleach>=6.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
Pygments>=2.15.0
Pillow>=10.0.0
aiofiles>=23.2.0
httpx>=0.24.0  # For OAuth and external API calls

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
